chore: only require deployment approval for 3P contributors
Mechanism:

We now have 2 otherwise identical environments, 'IntegTestCredentials'
and 'IntegTestCredentialsRequireApproval', except one of them requires
approvals to run.

A job will decide between them based on the submitter of the PR.
If the submitter is a project member or our bot, we will pick the one
that doesn't require approvals.

The output of this job is used as the value of the environment to use.