chore: "skip approval" check failing for some members
For some reason, @mrgrain is being marked as a `CONTRIBUTOR` (as opposed
to a `MEMBER`) for this repository. This means he is not eligible for
getting this PRs integ-tested without human approval.

@RomainMuller had the insight that we could simplify the check to saying
"is the PR from a fork or not". Presumably people with write access are
trusted, people who submit a forked PR are not.

Let's try it like this.