fix: support new Maven central publishing method
Maven Central is deprecating its old Nexus 2 OSSRH endpoint, replacing
it with a new plugin and protocol, and a compatibility endpoint.

I just want to migrate to the new publishing mode
-------------------------------------------------

To migrate, do the following:

* Configure `MAVEN_SERVER_ID=central-ossrh`,
* Log in to <https://central.sonatype.com/>
* Generate a new username and password on the new publisher using the **Generate User Token** feature.
* Configure `MAVEN_STAGING_PROFILE_ID` with your package namespace (register it first if necessary)

I'm going to review this PR
---------------------------

Because the new plugin doesn't support our 2-stage publishing method
yet, we are using the compatibility endpoint. The compatibility endpoint
works mostly with the `nexus-staging-maven-plugin` for uploading jars,
but the "release" command doesn't work because of a protocol
incompatibility. So instead, we are using a new custom HTTP request for
the very final "release" command, which is a slightly different version
of what the Nexus plugin sends, but *is* accepted by the endpoint.

Because the old script was in bash and hard to understand and modify,
I've rewritten it to TypeScript with zx, to at least get better type
checking and parameter modeling.

Also described the various parameters better in the README.