fix(publib-maven): fails with "Subprocess failed with exit code 1" (#1677)
In the old bash version, we had the following command:

```sh
set -eu

some_function() {
    export GPG_TTY=$(tty)
}
```

If there is no tty, this fails with exit code 1 but that doesn't stop
the script if the failing subshell is part of an `export` command.

On GitHub Actions there is no tty which didn't stop the original script,
but *does* stop the new script.

Configure `nothrow: true` on the `tty` subshell to avoid stopping the
command.